Revert "Rename models.rs to query_plan.rs"
This reverts commit e88e406a143694821b45cc04ebaa511bd581fd39.
